ml/requirements.txt
scikit-learn==1.5.2
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pg8000==1.31.2
python-dotenv==1.0.1
joblib==1.4.2
numpy==1.26.4
